create Sourced class
