+++ gamedata/compendium.py
@@ -5,12 +5,11 @@ import os
 
 import newrelic.agent
 
-from cogs5e.models.background import Background
 from cogs5e.models.monster import Monster
-from cogs5e.models.race import Race
 from cogs5e.models.spell import Spell
+from gamedata.background import Background
+from gamedata.race import Race
 from utils import config
-from utils.functions import parse_data_entry
 
 log = logging.getLogger(__name__)
